Move to api/views/emails.py
